audio-text-transcriber/requirements.txt
openai-whisper
ffmpeg
numpy
pydub
audiop